Fixed flake8 problems, added setup.py
Can install with pip now
