+++ setup.py
@@ -1,0 +1,20 @@
+from setuptools import setup
+from os import path
+
+
+here = path.abspath(path.dirname(__file__))
+print(here)
+
+setup(
+    name='UnDockerize',
+    version='0.0.1',
+    description='Creates an ansible file from a Dockerfile',
+    long_description='Creates an ansible file from a Dockerfile',
+    url='https://github.com/mellena1/UnDockerize',
+    author='Andrew Mellen',
+    author_email='andrew_mellen@icloud.com',
+    license='MIT',
+    entry_points={
+        'console_scripts': ['undockerize = undockerize.undockerize:main']
+    }
+)
